minor edits in Info.plist
Deleted references to the Avestan script (Avst) and to the present-day
country of Iran (IR). The first should be obvious, and now the keyboard
is simply named “Avestan” in the list of available keyboards, instead
of “Avestan (Avestan)”. I thought this would be clearer to the user.
The deletion of the country designation “IR” suggested itself because
the Avestan language and script are frequently used outside of Iran.